refactor: move mobile menu into own file
