+++ site/src/layout/header/MobileMenu.fragment.ts
@@ -1,0 +1,16 @@
+import { gql } from "@comet/cms-site";
+
+export const mobileMenuFragment = gql`
+    fragment MobileMenu on PageTreeNode {
+        id
+        name
+        hideInMenu
+        ...PageLink
+        childNodes {
+            id
+            name
+            hideInMenu
+            ...PageLink
+        }
+    }
+`;
